# improve base uri for ctx