--- a/ctx_test.go
+++ b/ctx_test.go
@@ -259,7 +259,7 @@ func Test_Ctx_BaseURL(t *testing.T) {
 	utils.AssertEqual(t, "http://google.com", c.BaseURL())
 }
 
-// go test -v -run=^$ -bench=Benchmark_Ctx_Append -benchmem -count=4
+// go test -v -run=^$ -bench=Benchmark_Ctx_BaseURL -benchmem
 func Benchmark_Ctx_BaseURL(b *testing.B) {
 	app := New()
 	c := app.AcquireCtx(&fasthttp.RequestCtx{})