bench compare result:
```
benchmark                    old ns/op     new ns/op     delta
Benchmark_Ctx_BaseURL-12     58.0          2.24          -96.14%

benchmark                    old allocs     new allocs     delta
Benchmark_Ctx_BaseURL-12     1              0              -100.00%

benchmark                    old bytes     new bytes     delta
Benchmark_Ctx_BaseURL-12     32            0             -100.00%
```